Fix the enum and added the dragon knight picture

+++ src/hom2/gamelogic/Characters/ObserverMinion.java
@@ -1,5 +1,5 @@
 package hom2.gamelogic.Characters;
 
 public interface ObserverMinion {
-	public void update(int hp, int minionAP);
+	public void update(long hp, long minionAP);
 }
